requirements.txt
asgiref==3.3.1
colorama==0.4.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.6
django-ajax-selects==2.1.0
django-import-export==2.5.0
django-qr-code==2.1.0
et-xmlfile==1.0.1
img2pdf==0.4.0
jdcal==1.4.1
lml==0.1.0
lxml==4.6.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.6
pikepdf==2.5.2
Pillow==8.1.0
psycopg2==2.8.6
pyexcel-io==0.6.4
pyexcel-xls==0.6.2
pyexcel-xlsx==0.6.0
pypng==0.0.20
PyQRCode==1.2.1
pytz==2021.1
PyYAML==5.4.1
qrcode==6.1
segno==1.3.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
xlrd==1.2.0
xlwt==1.3.0
